Move common logic into CAS1TimelineService
This is in preparation for changes to the timeline response
